Moved license command and queries to billing
