+++ src/Core/Billing/OrganizationFeatures/OrganizationLicenses/Interfaces/IUpdateOrganizationLicenseCommand.cs
@@ -4,7 +4,7 @@ using Bit.Core.AdminConsole.Entities;
 using Bit.Core.Billing.Models.Business;
 using Bit.Core.Models.Data.Organizations;
 
-namespace Bit.Core.OrganizationFeatures.OrganizationLicenses.Interfaces;
+namespace Bit.Core.Billing.OrganizationFeatures.OrganizationLicenses.Interfaces;
 
 public interface IUpdateOrganizationLicenseCommand
 {
