+++ test/Core.Test/Billing/OrganizationFeatures/OrganizationLicenses/UpdateOrganizationLicenseCommandTests.cs
@@ -1,10 +1,10 @@
 ﻿using System.Security.Claims;
 using Bit.Core.AdminConsole.Entities;
 using Bit.Core.Billing.Models.Business;
+using Bit.Core.Billing.OrganizationFeatures.OrganizationLicenses;
 using Bit.Core.Billing.Services;
 using Bit.Core.Enums;
 using Bit.Core.Models.Data.Organizations;
-using Bit.Core.OrganizationFeatures.OrganizationLicenses;
 using Bit.Core.Services;
 using Bit.Core.Settings;
 using Bit.Test.Common.AutoFixture;
@@ -14,7 +14,7 @@ using NSubstitute;
 using Xunit;
 using JsonSerializer = System.Text.Json.JsonSerializer;
 
-namespace Bit.Core.Test.OrganizationFeatures.OrganizationLicenses;
+namespace Bit.Core.Test.Billing.OrganizationFeatures.OrganizationLicenses;
 
 [SutProviderCustomize]
 public class UpdateOrganizationLicenseCommandTests
